# api/src/config/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Swagger Accounts
servers:
  - url: http://localhost:8080/api/v1
paths:
  /accounts:
    get:
      tags:
        - Account
      summary: Get all accounts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetAccount"
    post:
      tags:
        - Account
      summary: Add a new account
      description: ""
      operationId: addAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccount"
        description: The account to be added
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/GetAccount"
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  "/accounts/{id}":
    get:
      tags:
        - Account
      summary: Find account by ID
      description: Returns as single account
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          description: Id of account to return
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/GetAccount"
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
    put:
      tags:
        - Account
      summary: Update existing account
      operationId: updateAccount
      parameters:
        - name: id
          in: path
          description: Id of account to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccount"
        description: The account to be updated
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/GetAccount"
        "400":
          description: Bad request
        "500":
          description: Internal server message
    delete:
      tags:
        - Account
      summary: Delete existing account
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          description: Id of account to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  "/accounts/{id}/tag":
    post:
      tags:
        - Account Tag
      summary: add tag to an account
      parameters:
        - name: id
          in: path
          description: Id of account to add tag
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tag"
        description: The tag to be added
        required: true
      responses:
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
    delete:
      tags:
        - Account Tag
      summary: delete tag from an account
      parameters:
        - name: id
          in: path
          description: Id of account to delete tag from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tag"
        description: The tag to be deleted
        required: true
      responses:
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
components:
  schemas:
    UpdateAccount:
      type: object
      properties:
        name_first:
          type: string
        name_last:
          type: string
        employer:
          type: string
        phone:
          type: string
          pattern: '^\d{10}$'
        email:
          type: string
          format: email
        address:
          type: string
        picture:
          type: string
          format: uri
        balance:
          type: number
        credit:
          type: number
        comments:
          type: string
    CreateAccount:
      type: object
      properties:
        name_first:
          type: string
        name_last:
          type: string
        employer:
          type: string
        phone:
          type: string
          pattern: '^\d{10}$'
        email:
          type: string
          format: email
        address:
          type: string
        picture:
          type: string
          format: uri
        balance:
          type: number
        credit:
          type: number
        comments:
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        required:
          - name_first
          - name_last
          - employer
          - phone
          - email
          - address
          - picture
          - balance
          - credit
    GetAccount:
      allOf:
        - $ref: '#/components/schemas/CreateAccount'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            tags:
              type: array
              items:
                $ref: "#/components/schemas/Tag"
            created:
              type: string
              format: date-time

    Tag:
      type: object
      required:
        - name
      properties:
        name:
          type: string